Avoid compiler errors
... at least using MSVC 2019.

+++ src/deadbeef_rand.c
@@ -23,5 +23,5 @@ uint32_t deadbeef_generate_seed(void)
 {
 	  uint32_t t = (uint32_t)time(NULL);
 	  uint32_t c = (uint32_t)clock();
-	  return (t << 24) ^ (c << 11) ^ t ^ (size_t) &c;
+	  return (uint32_t)((t << 24) ^ (c << 11) ^ t ^ (size_t) &c);
 }
